refactor(web): extract long_tail_console.js from long_tail.js
web/js/long_tail.js (2436 lines) mixed the list shell (band tabs, search,
row rendering) with the per-row action console (U4/U5/U6, the #481 item 1
ConsoleState consolidation). #522 splits the console into its own module;
long_tail.js keeps the shell and imports consolePrune/consoleStates/
restoreLongTailConsoles back from the console module (a deliberate,
runtime-only ES-module cycle — every cross-reference is inside a function
body, never at module top-level, so node --check and the mjs suite both
confirm no TDZ error at import). main.js and pipeline.js's imports split
accordingly; test_js_long_tail_console.mjs now imports from the console
module and test_js_util.mjs destructures console helpers from a second
__test__ import.

Also folds in the two Workstream D cosmetic fixes:
- Repoints refetchLongTailRow's stale {@link markRowInFlight} JSDoc
  reference (the symbol no longer exists) to describe the actual
  double-fire-guard mechanism.
- checkYoutube now checks for a missing mb_release_id BEFORE touching
  consoleStates at all, so a row with no release identifier never leaves
  behind a residual, empty ConsoleState entry that would otherwise sit in
  the map until the next consolePrune. New coverage in
  test_js_long_tail_console.mjs.
