+++ web/js/long_tail.js
@@ -1,140 +1,45 @@
 // @ts-check
 
 /**
- * Long-tail triage worklist (Pipeline sub-view).
+ * Long-tail triage worklist (Pipeline sub-view) — the list shell.
  *
- * U3: the list shell — band tabs (with live counts), a search box, and
- * the row list. Fetches `GET /api/pipeline/long-tail` once (KTD2 — one
- * server-banded fetch, all tab/search filtering happens client-side over
- * that single payload), derives the band tab set from the bands present
- * in the cohort, and renders the rows for the selected band.
+ * U3: band tabs (with live counts), a search box, and the row list. Fetches
+ * `GET /api/pipeline/long-tail` once (KTD2 — one server-banded fetch, all
+ * tab/search filtering happens client-side over that single payload),
+ * derives the band tab set from the bands present in the cohort, and
+ * renders the rows for the selected band.
  *
- * U4 (this unit): the in-place action console. Selecting a row expands
- * its `.lt-detail` container into a band-aware evidence console with five
- * panels, each loading INDEPENDENTLY (per-panel loading + error states)
- * so a slow / failing `GET /api/release-group/<rg>` (the 15s MB-mirror
- * timeout) never blanks the console or silently drops another panel:
- *
- *   1. Why-unfindable  ← `GET /api/triage/<id>` (unfindable category +
- *      reason + search-forensics rollup; "not yet categorised" rendered
- *      distinctly from an error).
- *   2. Soulseek peers  ← `GET /api/pipeline/<id>` `last_search.top_candidates`
- *      (reuses `renderForensicBlock`; a few rows with a "show all" toggle).
- *   3. Recent rescues  ← `GET /api/pipeline/<id>` `history` rows where
- *      `source==='youtube'` ("rescue running" / "last rescue failed: …").
- *   4. Sibling pressings ← `GET /api/release-group/<rg>` (rg read straight
- *      off the cohort row's `mb_release_group_id` — projected by
- *      `_LONG_TAIL_SELECT`, #398; skipped only for rows with no release
- *      group / master at all. `mb_release_group_id` doubles as the
- *      Discogs master id for Discogs-sourced rows (KTD-1), and the route
- *      dispatches numeric ids to the Discogs master endpoint server-side
- *      (#501 item 1) — a Discogs row with a resolved master gets real
- *      sibling pressings here too, not a doomed MB lookup).
- *   5. YouTube matrix  ← the four-state shell. Defaults to `never_run`
- *      with a "Check YouTube" button (U5 wires the actual resolver call —
- *      U4 must NOT auto-call the slow, side-effectful resolver GET).
- *
- * U5 (this unit): the two-step YouTube rescue flow.
- *
- *   1. "Check YouTube" (`checkYoutube`) — replaces U4's placeholder. Calls
- *      the SLOW, SIDE-EFFECTFUL resolver GET
- *      (`GET /api/youtube-album?identifier=<mb_release_id>`), disables the
- *      button + shows in-progress, GUARDS double-fire
- *      (`consoleCanStart(consoleStates, id, 'resolve')` — a second click
- *      while outstanding fires nothing), and STAMPS the fetch with a per-row
- *      console token so a stale result (operator collapsed the console or
- *      moved to another row) never paints. On return the YouTube panel is
- *      re-rendered via `youtubeSectionState`: `resolved_with_matrix` →
- *      pickable rescue targets; `resolved_empty` → "not on YouTube Music —
- *      re-check" (rescue affordance HIDDEN); `resolver_failed` → error +
- *      Retry; cached-but-stale → matrix with a "cached" note.
- *   2. Pick + confirm + submit — clicking a matrix target's "Rescue from
- *      this" (`pickYoutubeRescue`) opens a Promise-based confirm overlay
- *      (mirrors `web/js/replace_picker.js`'s `.confirm-overlay` shell +
- *      backdrop-click-cancel). On confirm →
- *      `POST /api/pipeline/<id>/youtube-rescue {browse_id}`. Every ingest
- *      outcome maps to specific console copy (`rescueOutcomeCopy`). On
- *      `accepted` the row is marked in-flight and a SINGLE-ROW refetch
- *      (`GET /api/pipeline/long-tail?id=<id>`) patches just that cohort
- *      row (KTD8 — never optimistically move the row to a band; the
- *      importer owns the transition, KTD4).
- *
- * U6 (this unit): the console's secondary operator actions, plus the
- * shared post-action freshness model.
- *
- *   1. Set-imported — accepts the on-disk quality: flips the request
- *      `wanted → imported` via the existing `POST /api/pipeline/update`
- *      surface, then drops just that row from the cohort.
- *   2. Accept-a-sibling-pressing — wires the existing Replace operator
- *      action (`openReplacePicker`, standard mode) so the operator can
- *      switch the request to a different pressing in the same MB release
- *      group. MB-only (KTD7): disabled with a one-line reason for
- *      Discogs-sourced requests or rows with no `mb_release_group_id`
- *      (the `canAcceptSibling` predicate). On a confirmed Replace the old
- *      request becomes a frozen `replaced` row that LEAVES the worklist —
- *      so this is the one action that closes the console and does a FULL
- *      cohort refetch (`loadLongTail`), not the single-row patch.
- *   3. Set-intent — renders the request's current quality intent
- *      (lossless vs. default, from `target_format`) as a badge with a
- *      toggle; switching posts `{id, intent}` to the existing set-intent
- *      surface (`intentToggleTarget` picks the opposite of the current
- *      intent), then single-row refetch-and-patch so the badge reflects
- *      the new value.
- *   4. Delete-request — removes the request entirely via the existing
- *      `POST /api/pipeline/delete` surface (confirm-gated; a superseding
- *      Replace blocks it with a 409 + descendant ids), then drops the row.
- *
- * Shared freshness (KTD8): the set-intent action awaits then refetches just
- * the acted-on row (`refetchLongTailRow`, the U5 helper) and patches it in —
- * no optimistic band moves, no polling. Set-imported / Delete remove just
- * the acted-on row's DOM nodes (it leaves the cohort); the confirmed Replace
- * closes the console and full-cohort refetches. Open consoles survive every
- * list re-render: `consoleStates` tracks each row's open flag and
- * `restoreLongTailConsoles` re-opens them after the DOM wipe (#398).
- *
- * #481 item 1 (this unit): the console-state consolidation. Every per-row
- * bookkeeping structure the console needs — the fetch token, the five
- * action double-fire guards, the cached YouTube resolver result, and the
- * open/closed flag — used to live in eight parallel module-scoped
- * structures (plus `state.longTail.open` on shared state) with two
- * independent prune sites that each had to remember to sweep every one of
- * them (the #480 review caught two "forgot to prune one" instances — an
- * open-set leak on full refetch, a stale resolver cache). `consoleStates`
- * (a single `Map<id, ConsoleState>`) replaces all eight; five pure
- * transition helpers (`consoleOpen`/`consoleClose`/`consolePrune`/
- * `consoleCanStart`/`consoleSettle`) are the only way to mutate it, and
- * `consolePrune` is the ONE prune function called from both former prune
- * sites. The two deliberate token semantics from U5 are now explicit,
- * separately named functions rather than comment-only: `consoleIsStale`
- * (panel-paint — discard against the token CAPTURED at fetch time) vs.
- * `consoleToken` (resolver-settle — read the CURRENT token at paint time).
+ * The in-place per-row action console (U4/U5/U6, and the #481 item 1
+ * `ConsoleState` consolidation) lives in `./long_tail_console.js` — split
+ * out by #522. The two modules share a deliberate, SAFE runtime-only
+ * ES-module cycle: `loadLongTail` (below) prunes `consoleStates` via
+ * `consolePrune` when a row leaves the cohort, and `onLongTailSearchInput`
+ * (below) restores expanded consoles via `restoreLongTailConsoles` after a
+ * search repaint wipes the list DOM. Every cross-reference is inside a
+ * function body, never at module top-level, so neither module's top-level
+ * evaluation can observe the other mid-load — `node --check` and the mjs
+ * test suite both confirm no TDZ error at import.
  *
  * Pure / DOM-free helpers (band ordering, tab derivation, in-band search
- * filtering, cross-band match count, YouTube state classifier, console
- * emphasis selector, and the #481 console-state transitions) are exported
- * via `__test__` for the Node unit suite. The classifier + emphasis
- * selector live in `util.js` (the shared pure home) and are re-exported
- * through `__test__` here for convenience. Rendering and fetch live
- * alongside them but never leak into the pure helpers.
+ * filtering, cross-band match count) are exported via `__test__` for the
+ * Node unit suite. Rendering and fetch live alongside them but never leak
+ * into the pure helpers.
  *
  * Shape mirrors `web/js/search_plan.js` / `web/js/recents.js`:
  * `// @ts-check`, ES6 module, JSDoc on exports, the
  * `export const __test__ = {…}` named-object test convention.
  */
 
-import { state, API, toast } from './state.js';
+import { state, API } from './state.js';
 import {
   esc,
-  jsArg,
-  renderForensicBlock,
-  youtubeSectionState,
-  consoleEmphasis,
-  overrideToIntent,
-  awstDateTime,
   sourceLabel,
-  youtubeBrowseUrl,
 } from './util.js';
-import { openReplacePicker, pressingMeta } from './replace_picker.js';
+import {
+  consolePrune,
+  consoleStates,
+  restoreLongTailConsoles,
+} from './long_tail_console.js';
 
 /**
  * The `Missing` band sentinel — a `wanted` request with no clean
@@ -619,1771 +524,6 @@ export function onLongTailSearchInput(value) {
   }
 }
 
-// --- Action console (U4) --------------------------------------------
-//
-// Five evidence panels rendered in place, each loading independently. The
-// pure render helpers below take already-fetched data (or an error
-// sentinel) and return an HTML string; the DOM-side loaders fetch each
-// source and patch only that panel's container. A per-row token guard
-// discards a console fetch that resolves after the operator has clicked a
-// different row (or re-collapsed this one).
-//
-// #481 item 1: all per-row console bookkeeping — the fetch token, the
-// five action double-fire guards (resolve/submit/intent/import/delete),
-// the cached YouTube resolver result, and the open/closed flag — used to
-// live in eight parallel module-scoped structures (plus `state.longTail.
-// open` on shared state) with two independent prune sites that each had to
-// remember to sweep every one of them. The #480 review caught two
-// "forgot to prune one" instances that shape produced. `consoleStates`
-// below is the single `Map<id, ConsoleState>` replacing all eight; the
-// five pure transition helpers (`consoleOpen`/`consoleClose`/
-// `consolePrune`/`consoleCanStart`/`consoleSettle`) are the only way to
-// mutate it, and `consolePrune` is the ONE prune function called from both
-// sites (`removeRowFromCohort`, `loadLongTail`'s fresh-cohort intersect).
-
-/**
- * Per-row console state — the single source of truth for everything the
- * action console needs to track per `album_requests.id`. Replaces the
- * eight parallel structures #481 item 1 called out (fetch token, five
- * in-flight guard Sets, the YouTube-result cache, and the open-console
- * Set) with one `Map<id, ConsoleState>` entry per row.
- *
- * @typedef {Object} ConsoleState
- * @property {boolean} open  Is this row's console currently expanded?
- * @property {number} token  Per-row console-fetch token, bumped on every
- *   {@link consoleOpen} / {@link consoleClose} so a panel fetch stamped
- *   with a stale token is discarded before it paints ({@link consoleIsStale}).
- * @property {Object|null} youtubeResult  The row's cached YouTube resolver
- *   result (from a prior Check YouTube this session), so a console restore
- *   after a list re-render doesn't reset a resolved matrix back to
- *   `never_run` (#398).
- * @property {Set<string>} inFlight  Action names (`'resolve'`, `'submit'`,
- *   `'intent'`, `'import'`, `'delete'`) with an outstanding request —
- *   double-fire guards, gated by {@link consoleCanStart} /
- *   {@link consoleSettle}.
- */
-
-/**
- * Per-row console state, keyed by `album_requests.id`. The single map
- * backing every pure transition helper below — see the header comment for
- * the eight-structures-to-one consolidation this replaces.
- *
- * @type {Map<number, ConsoleState>}
- */
-const consoleStates = new Map();
-
-/**
- * Get (creating if absent) the console-state entry for `id`. Every mutator
- * below routes through this so a row's first touch — whichever action
- * fires first — always finds a well-formed entry. Pure.
- *
- * @param {Map<number, ConsoleState>} map
- * @param {number} id
- * @returns {ConsoleState}
- */
-function ensureConsoleState(map, id) {
-  let entry = map.get(id);
-  if (!entry) {
-    entry = { open: false, token: 0, youtubeResult: null, inFlight: new Set() };
-    map.set(id, entry);
-  }
-  return entry;
-}
-
-/**
- * Open (or re-open) a row's console. Bumps the token so a panel fetch fired
- * against a stale console (operator collapsed/reopened while it was
- * outstanding) is discarded on resolve, and marks the row open. Pure.
- *
- * @param {Map<number, ConsoleState>} map
- * @param {number} id
- * @returns {number} The new token — stamp the panel fetches this open fires.
- */
-export function consoleOpen(map, id) {
-  const entry = ensureConsoleState(map, id);
-  entry.open = true;
-  entry.token += 1;
-  return entry.token;
-}
-
-/**
- * Close a row's console. Bumps the token (same discard effect as
- * {@link consoleOpen}) and marks the row closed, but deliberately leaves
- * `inFlight` and `youtubeResult` untouched: an outstanding action keeps its
- * double-fire guard across a collapse (LT-R1 is a SEPARATE, explicit clear —
- * see {@link consoleClearGuards}), and the cached resolver result survives so
- * a later reopen restores the matrix instead of resetting to `never_run`
- * (#398). Pure.
- *
- * @param {Map<number, ConsoleState>} map
- * @param {number} id
- * @returns {number} The new token.
- */
-export function consoleClose(map, id) {
-  const entry = ensureConsoleState(map, id);
-  entry.open = false;
-  entry.token += 1;
-  return entry.token;
-}
-
-/**
- * Drop every row's console state whose id does NOT satisfy `keep`. The
- * ONE prune function called from BOTH prune sites — `loadLongTail`'s
- * fresh-cohort intersect (`keep = (id) => cohortIds.has(id)`) and
- * `removeRowFromCohort`'s single-row drop (`keep = (id) => id !== removedId`).
- * Removing an id drops its ENTIRE row state atomically — open flag, token,
- * cached YouTube result, in-flight guards — in one call, so there is no
- * longer a second (or third…) structure a future change can forget to sweep
- * (#481 item 1). Pure.
- *
- * @param {Map<number, ConsoleState>} map
- * @param {(id: number) => boolean} keep
- * @returns {void}
- */
-export function consolePrune(map, keep) {
-  for (const id of [...map.keys()]) {
-    if (!keep(id)) map.delete(id);
-  }
-}
-
-/**
- * Atomic check-and-set double-fire guard for one row's named action
- * (`'resolve'` | `'submit'` | `'intent'` | `'import'` | `'delete'`). Returns
- * `true` and marks the action in-flight when none was outstanding for this
- * id; returns `false` (no mutation) when one already was. Pair with
- * {@link consoleSettle} in a `try/finally` around the async call. Pure.
- *
- * @param {Map<number, ConsoleState>} map
- * @param {number} id
- * @param {string} action
- * @returns {boolean}
- */
-export function consoleCanStart(map, id, action) {
-  const entry = ensureConsoleState(map, id);
-  if (entry.inFlight.has(action)) return false;
-  entry.inFlight.add(action);
-  return true;
-}
-
-/**
- * Clear one row's named action in-flight flag — the `finally` counterpart
- * of {@link consoleCanStart}. No-op when the row has no tracked state
- * (already pruned — e.g. the row left the cohort while its fetch was
- * outstanding). Pure.
- *
- * @param {Map<number, ConsoleState>} map
- * @param {number} id
- * @param {string} action
- * @returns {void}
- */
-export function consoleSettle(map, id, action) {
-  const entry = map.get(id);
-  if (entry) entry.inFlight.delete(action);
-}
-
-/**
- * Clear EVERY in-flight action guard for one row (LT-R1) — called when the
- * operator (re)opens or collapses THIS row's console via
- * {@link toggleLongTailDetail}, so a console destroyed mid-flight doesn't
- * leave a stale guard silently dead-ing the fresh console's buttons until an
- * orphaned fetch settles. Deliberately NOT called by
- * {@link restoreLongTailConsoles} — an outstanding operation keeps its
- * double-fire protection across a list re-render (#398 fidelity). Pure.
- *
- * @param {Map<number, ConsoleState>} map
- * @param {number} id
- * @returns {void}
- */
-export function consoleClearGuards(map, id) {
-  const entry = map.get(id);
-  if (entry) entry.inFlight.clear();
-}
-
-/**
- * The row's CURRENT console token — the resolver-settle semantic. Used to
- * paint the YouTube panel against whatever console exists NOW (even one
- * re-created by a #398 restore while the resolver fetch was outstanding),
- * never the token captured when the fetch fired. Contrast
- * {@link consoleIsStale}, the panel-paint semantic, which discards against
- * the CAPTURED token instead. Returns `0` for an untracked id (never
- * opened, or already pruned). Pure.
- *
- * @param {Map<number, ConsoleState>} map
- * @param {number} id
- * @returns {number}
- */
-export function consoleToken(map, id) {
-  const entry = map.get(id);
-  return entry ? entry.token : 0;
-}
-
-/**
- * Stale-discard predicate for a panel fetch stamped with `capturedToken` at
- * fire time — the panel-paint semantic (contrast {@link consoleToken}, the
- * resolver-settle semantic). `true` means the console has moved on
- * (reopened, closed, or pruned) since the fetch fired, and the result must
- * be discarded rather than painted. Pure.
- *
- * @param {Map<number, ConsoleState>} map
- * @param {number} id
- * @param {number} capturedToken
- * @returns {boolean}
- */
-export function consoleIsStale(map, id, capturedToken) {
-  return consoleToken(map, id) !== capturedToken;
-}
-
-/**
- * Is this row's console currently marked open? Pure.
- *
- * @param {Map<number, ConsoleState>} map
- * @param {number} id
- * @returns {boolean}
- */
-export function consoleIsOpen(map, id) {
-  const entry = map.get(id);
-  return entry != null && entry.open === true;
-}
-
-/**
- * The row's cached YouTube resolver result, or `null` if none settled yet.
- * Pure.
- *
- * @param {Map<number, ConsoleState>} map
- * @param {number} id
- * @returns {Object|null}
- */
-export function consoleYoutubeResult(map, id) {
-  const entry = map.get(id);
-  return entry ? entry.youtubeResult : null;
-}
-
-/**
- * Cache a row's settled YouTube resolver result (#398 restore cache). Pure.
- *
- * @param {Map<number, ConsoleState>} map
- * @param {number} id
- * @param {Object|null} result
- * @returns {void}
- */
-export function consoleSetYoutubeResult(map, id, result) {
-  ensureConsoleState(map, id).youtubeResult = result;
-}
-
-/**
- * Every id currently marked open, in insertion order — drives
- * {@link restoreLongTailConsoles}. Pure.
- *
- * @param {Map<number, ConsoleState>} map
- * @returns {number[]}
- */
-export function consoleOpenIds(map) {
-  const ids = [];
-  for (const [id, entry] of map) {
-    if (entry.open) ids.push(id);
-  }
-  return ids;
-}
-
-/**
- * Visible-row cap for the Soulseek peers panel before the "show all"
- * expansion. The peers list shows a few rows so the action buttons (U5/U6)
- * stay reachable without scrolling past a long candidate table.
- *
- * @type {number}
- */
-const PEERS_VISIBLE_CAP = 5;
-
-/**
- * Format an ISO timestamp for the console, defensively. Falls back to the
- * raw string when it can't be parsed (never throws into a render).
- *
- * @param {string|null|undefined} iso
- * @returns {string}
- */
-function consoleTimestamp(iso) {
-  if (!iso) return '';
-  try {
-    return awstDateTime(String(iso));
-  } catch (_e) {
-    return String(iso);
-  }
-}
-
-/**
- * A small panel scaffold: a titled section with a body. Used by every
- * panel so the console reads consistently. Pure.
- *
- * @param {string} name   Stable slug for the panel container id.
- * @param {number} id     album_requests.id (namespaces the container).
- * @param {string} title  Panel heading.
- * @param {string} body   Inner HTML (already escaped where needed).
- * @param {boolean} [lead]  When true, flag the panel as the console's
- *   lead panel (band-aware emphasis).
- * @returns {string}
- */
-function renderPanel(name, id, title, body, lead) {
-  const leadCls = lead ? ' lt-panel-lead' : '';
-  return `<div class="lt-panel lt-panel-${esc(name)}${leadCls}" id="lt-panel-${esc(name)}-${id}">
-    <div class="lt-panel-title">${esc(title)}</div>
-    <div class="lt-panel-body">${body}</div>
-  </div>`;
-}
-
-/**
- * Per-panel loading affordance (distinct from empty/error). Pure.
- *
- * @param {string} label
- * @returns {string}
- */
-function renderPanelLoading(label) {
-  return `<div class="lt-panel-loading">Loading ${esc(label)}…</div>`;
-}
-
-/**
- * Per-panel error affordance — the independent-load contract: one panel's
- * fetch failing renders THIS, never blanking the console or dropping
- * another panel. Pure.
- *
- * @param {string} label
- * @returns {string}
- */
-function renderPanelError(label) {
-  return `<div class="lt-panel-error">Couldn't load ${esc(label)}. <span class="lt-panel-error-hint">(other panels are unaffected)</span></div>`;
-}
-
-/**
- * Render the why-unfindable panel body from a triage payload. Pure.
- *
- * Three states:
- *   * `triage == null` → caller passed nothing yet (defensive; the loader
- *     uses the loading/error affordances instead).
- *   * `triage.unfindable == null` → NOT yet categorised. Rendered as an
- *     explicit "detection runs daily" state — not an error, not blank
- *     (R7).
- *   * categorised → the category + a search-forensics rollup.
- *
- * @param {Object|null} triage  The `TriageResult` payload.
- * @returns {string}
- */
-function renderUnfindableBody(triage) {
-  if (!triage || typeof triage !== 'object') {
-    return '<div class="lt-panel-empty">No triage data.</div>';
-  }
-  const unfindable = triage.unfindable;
-  const sf = triage.search_forensics || {};
-  const rollupParts = [];
-  if (sf.total_searches != null) rollupParts.push(`${sf.total_searches} searches`);
-  if (sf.with_cands_count != null) rollupParts.push(`${sf.with_cands_count} with candidates`);
-  if (sf.zero_results_count != null) rollupParts.push(`${sf.zero_results_count} zero-result`);
-  if (sf.dominant_rejection_reason) rollupParts.push(`dominant reject: ${sf.dominant_rejection_reason}`);
-  const lastAt = sf.last_search_at ? consoleTimestamp(sf.last_search_at) : '';
-  const rollup = rollupParts.length
-    ? `<div class="lt-rollup">${esc(rollupParts.join(' · '))}${lastAt ? ` · last ${esc(lastAt)}` : ''}</div>`
-    : '';
-  if (!unfindable || unfindable.category == null) {
-    // Not-yet-categorised — daily detection state, distinct from an error.
-    return `<div class="lt-uncategorised">
-      <span class="lt-meta-chip">not yet categorised</span>
-      <span class="lt-uncategorised-note">detection runs daily</span>
-    </div>${rollup}`;
-  }
-  const catAt = unfindable.categorised_at ? consoleTimestamp(unfindable.categorised_at) : '';
-  const probe = (unfindable.last_artist_probe_match_count != null)
-    ? `<div class="lt-rollup">artist probe: ${unfindable.last_artist_probe_match_count} match${unfindable.last_artist_probe_match_count === 1 ? '' : 'es'}${unfindable.last_artist_probe_at ? ` · ${esc(consoleTimestamp(unfindable.last_artist_probe_at))}` : ''}</div>`
-    : '';
-  return `<div class="lt-unfindable-cat">
-      <span class="badge badge-manual">${esc(String(unfindable.category))}</span>
-      ${catAt ? `<span class="lt-meta-chip">categorised ${esc(catAt)}</span>` : ''}
-    </div>${rollup}${probe}`;
-}
-
-/**
- * Pluck the YouTube rescue rows out of a pipeline-detail history list.
- * Pure. Returns the `source==='youtube'` rows newest-first (the history
- * already arrives newest-first; we preserve order).
- *
- * @param {Array<Object>|null|undefined} history
- * @returns {Array<Object>}
- */
-function youtubeHistoryRows(history) {
-  return (Array.isArray(history) ? history : [])
-    .filter((h) => h && h.source === 'youtube');
-}
-
-/**
- * Extract the classified failure reason for a terminal `youtube_failed`
- * download-history row. Pure. The reason is persisted in the
- * `youtube_metadata` JSONB blob (`reason`) by the ingest worker's terminal
- * write — see `lib/youtube_ingest_service.py::YoutubeIngestMetadata.reason`.
- * Falls back to the row's `error_message` then `verdict` (the shared
- * download-history-view fields) and finally a generic sentinel so the
- * panel never shows an empty reason.
- *
- * @param {Object} row  A `source==='youtube'` download-history row.
- * @returns {string}
- */
-function youtubeFailureReason(row) {
-  const meta = row && row.youtube_metadata;
-  if (meta && typeof meta === 'object' && meta.reason) {
-    return String(meta.reason);
-  }
-  if (row && row.error_message) return String(row.error_message);
-  if (row && row.verdict) return String(row.verdict);
-  return 'unknown';
-}
-
-/**
- * Render the "recent rescue attempts" panel body. Pure.
- *
- * Reads the `source==='youtube'` history rows (KTD4 — rescues never move
- * the request row; their state lives in `download_log`). Surfaces:
- *   * "rescue running" when an active `youtube_running` row exists, OR the
- *     worklist row carried `in_flight_rescue` (the same predicate).
- *   * "last rescue failed: <reason>" when the latest terminal youtube row
- *     is `youtube_failed` SPECIFICALLY (distinct from `youtube_success`).
- *   * the recent attempts list otherwise.
- *
- * @param {Array<Object>|null|undefined} history  Pipeline-detail history.
- * @param {boolean} inFlightFlag  The worklist row's `in_flight_rescue`.
- * @returns {string}
- */
-function renderRescuesBody(history, inFlightFlag) {
-  const rows = youtubeHistoryRows(history);
-  const running = rows.find((h) => h.outcome === 'youtube_running');
-  if (running || inFlightFlag) {
-    return `<div class="lt-rescue-status"><span class="badge badge-new">rescue running</span>${running && running.created_at ? `<span class="lt-meta-chip">since ${esc(consoleTimestamp(running.created_at))}</span>` : ''}</div>`;
-  }
-  // The latest TERMINAL youtube row (youtube_failed / youtube_success).
-  const terminal = rows.find(
-    (h) => h.outcome === 'youtube_failed' || h.outcome === 'youtube_success');
-  if (terminal && terminal.outcome === 'youtube_failed') {
-    const reason = youtubeFailureReason(terminal);
-    return `<div class="lt-rescue-status"><span class="badge badge-manual">last rescue failed</span> <span class="lt-rescue-reason">${esc(reason)}</span></div>`;
-  }
-  if (rows.length === 0) {
-    return '<div class="lt-panel-empty">No rescue attempts yet.</div>';
-  }
-  // Some succeeded / mixed — list the recent attempts.
-  const items = rows.slice(0, 5).map((h) => {
-    const when = h.created_at ? consoleTimestamp(h.created_at) : '';
-    return `<div class="lt-rescue-item">${esc(String(h.outcome || '?'))}${when ? ` · ${esc(when)}` : ''}</div>`;
-  }).join('');
-  return `<div class="lt-rescue-list">${items}</div>`;
-}
-
-/**
- * Render the Soulseek peers panel body. Reuses `renderForensicBlock` for
- * the candidates table; caps the visible rows with a "show all" toggle so
- * the action buttons stay reachable (IA per the plan). Pure.
- *
- * @param {Object|null|undefined} lastSearch  `last_search` payload.
- * @param {number} id  album_requests.id (namespaces the show-all toggle).
- * @returns {string}
- */
-function renderPeersBody(lastSearch, id) {
-  if (!lastSearch) {
-    return renderForensicBlock(null);
-  }
-  const cands = Array.isArray(lastSearch.top_candidates)
-    ? lastSearch.top_candidates : [];
-  if (cands.length <= PEERS_VISIBLE_CAP) {
-    return renderForensicBlock(/** @type {any} */ (lastSearch));
-  }
-  // Cap the visible candidates; offer a "show all" toggle that swaps in
-  // the full block. The capped + full blocks are both pre-rendered; the
-  // toggle flips which is shown without a refetch.
-  const capped = { ...lastSearch, top_candidates: cands.slice(0, PEERS_VISIBLE_CAP) };
-  const cappedHtml = renderForensicBlock(/** @type {any} */ (capped));
-  const fullHtml = renderForensicBlock(/** @type {any} */ (lastSearch));
-  return `<div class="lt-peers" id="lt-peers-${id}">
-    <div class="lt-peers-capped">${cappedHtml}
-      <button class="lt-link-btn" type="button" onclick="event.stopPropagation(); window.toggleLongTailPeers(${id})">show all ${cands.length} peers</button>
-    </div>
-    <div class="lt-peers-full" style="display:none;">${fullHtml}
-      <button class="lt-link-btn" type="button" onclick="event.stopPropagation(); window.toggleLongTailPeers(${id})">show fewer</button>
-    </div>
-  </div>`;
-}
-
-/**
- * Toggle the capped ⇄ full Soulseek peers view in place. No refetch — both
- * blocks are already rendered; this flips visibility.
- *
- * @param {number} id  album_requests.id
- * @returns {void}
- */
-export function toggleLongTailPeers(id) {
-  if (typeof document === 'undefined') return;
-  const wrap = document.getElementById(`lt-peers-${id}`);
-  if (!wrap) return;
-  const capped = /** @type {HTMLElement|null} */ (wrap.querySelector('.lt-peers-capped'));
-  const full = /** @type {HTMLElement|null} */ (wrap.querySelector('.lt-peers-full'));
-  if (!capped || !full) return;
-  const showFull = full.style.display === 'none';
-  full.style.display = showFull ? 'block' : 'none';
-  capped.style.display = showFull ? 'none' : 'block';
-}
-
-/**
- * Render one sibling-pressing row for the siblings panel. Pure.
- *
- * @param {Object} rel  A release row from `/api/release-group/<rg>`.
- * @returns {string}
- */
-function renderSiblingRow(rel) {
-  const title = rel.title || '?';
-  // Reuse the Replace picker's pressing-summary formatter so the siblings
-  // panel and the picker render the same "country · year · format · Nt".
-  const meta = pressingMeta(rel);
-  const inLib = rel.in_library
-    ? `<span class="badge badge-rank-${esc(String(rel.library_rank || 'library').toLowerCase())}">in library</span>`
-    : '';
-  const pStatus = rel.pipeline_status
-    ? `<span class="badge badge-${esc(String(rel.pipeline_status))}">${esc(String(rel.pipeline_status))}</span>`
-    : '';
-  return `<div class="lt-sibling">
-    <span class="lt-sibling-title">${esc(String(title))}</span>
-    <span class="lt-sibling-meta">${esc(meta)}</span>
-    ${inLib}${pStatus}
-  </div>`;
-}
-
-/**
- * Render the sibling-pressings panel body from a release-group payload.
- * Pure. Evidence-only in U4 — the accept-sibling action is U6.
- *
- * @param {Object|null|undefined} rgData  `{releases:[...]}`.
- * @returns {string}
- */
-function renderSiblingsBody(rgData) {
-  const releases = (rgData && Array.isArray(rgData.releases)) ? rgData.releases : [];
-  if (releases.length === 0) {
-    return '<div class="lt-panel-empty">No sibling pressings found.</div>';
-  }
-  return releases.map(renderSiblingRow).join('');
-}
-
-/**
- * @typedef {Object} YoutubeRescueTarget
- * @property {string} yt_browse_id  The YT Music album browseId — the value
- *   the rescue submit (`POST .../youtube-rescue`) takes as `browse_id`.
- * @property {number|null} year
- * @property {number|null} track_count
- * @property {number|null} best_distance  Lowest `ok` beets distance across
- *   the release's `distances[]`, or `null` when none scored.
- */
-
-/**
- * Lowest `ok` beets distance across a YT release's `distances[]`. Pure /
- * DOM-free. Returns `null` when no distance row scored `ok` (so callers can
- * render "no distance" without a special-case branch).
- *
- * @param {{distances?: Array<{outcome?: string, distance?: number}>|null}} rel
- * @returns {number|null}
- */
-export function youtubeBestDistance(rel) {
-  const dists = (rel && Array.isArray(rel.distances)) ? rel.distances : [];
-  let best = /** @type {number|null} */ (null);
-  for (const d of dists) {
-    if (!d || d.outcome !== 'ok' || typeof d.distance !== 'number') continue;
-    if (best == null || d.distance < best) best = d.distance;
-  }
-  return best;
-}
-
-/**
- * Extract the pickable rescue targets from a resolver result. Pure /
- * DOM-free. Each target carries its `yt_browse_id` (the value the rescue
- * submit needs) plus display meta (year / track_count / best distance).
- *
- * Only `resolved_with_matrix` yields targets — `resolved_empty` (ok with
- * zero releases), `resolver_failed`, and `never_run` (null) all yield an
- * empty list, so the rescue affordance is hidden (R9 / R10: the operator
- * picks a target, the system never auto-picks; "not on YouTube Music"
- * offers nothing to pick). A release missing a `yt_browse_id` is dropped —
- * it cannot be a rescue target without the id the submit requires.
- *
- * @param {{outcome?: string, youtube_releases?: Array<Object>|null, from_cache?: boolean, error_message?: string|null}|null|undefined} result
- * @returns {YoutubeRescueTarget[]}
- */
-export function youtubeRescueTargets(result) {
-  const cls = youtubeSectionState(result);
-  if (cls.state !== 'resolved_with_matrix') return [];
-  const releases = (result && Array.isArray(result.youtube_releases))
-    ? result.youtube_releases : [];
-  /** @type {YoutubeRescueTarget[]} */
-  const targets = [];
-  for (const rel of releases) {
-    const browseId = rel && rel.yt_browse_id ? String(rel.yt_browse_id) : '';
-    if (!browseId) continue;  // no id → not a pickable target.
-    targets.push({
-      yt_browse_id: browseId,
-      year: (rel.year != null) ? Number(rel.year) : null,
-      track_count: (rel.track_count != null) ? Number(rel.track_count) : null,
-      best_distance: youtubeBestDistance(rel),
-    });
-  }
-  return targets;
-}
-
-/**
- * Render the YouTube panel body for a given section state + payload. Pure.
- *
- * Renders all four states. In `resolved_with_matrix` each release is a
- * PICKABLE rescue target (U5): a "Rescue from this" button carrying the
- * release's `yt_browse_id` opens the confirm step
- * (`window.pickYoutubeRescue(id, browse_id)`). `resolved_empty` HIDES the
- * rescue affordance and shows "not on YouTube Music — re-check"
- * (R9 / R10 — nothing to pick). `never_run` / `resolver_failed` render the
- * "Check YouTube" / retry button wired to `window.checkYoutube(id)` (U5's
- * real resolver handler, replacing U4's placeholder). The console must NOT
- * auto-call the slow, side-effectful resolver GET — the panel opens in
- * `never_run` until the operator clicks.
- *
- * @param {{outcome?: string, youtube_releases?: Array<Object>|null, from_cache?: boolean, error_message?: string|null}|null} result
- *   A cached resolver result, or `null` for the default never-run state.
- * @param {number} id  album_requests.id
- * @returns {string}
- */
-function renderYoutubeBody(result, id) {
-  const cls = youtubeSectionState(result);
-  const checkLabel = (cls.state === 'resolver_failed') ? 'Retry'
-    : (cls.state === 'resolved_empty') ? 'Re-check'
-    : 'Check YouTube';
-  const checkBtn = `<button class="lt-yt-check" type="button" onclick="event.stopPropagation(); window.checkYoutube(${id})">${checkLabel}</button>`;
-  if (cls.state === 'never_run') {
-    return `<div class="lt-yt lt-yt-never-run">
-      <div class="lt-yt-prompt">Resolve this release against YouTube Music.</div>
-      ${checkBtn}
-    </div>`;
-  }
-  if (cls.state === 'resolver_failed') {
-    return `<div class="lt-yt lt-yt-failed">
-      <div class="lt-yt-msg">${esc(cls.message)}</div>
-      ${checkBtn}
-    </div>`;
-  }
-  if (cls.state === 'resolved_empty') {
-    // Not on YouTube Music — HIDE the rescue affordance (nothing to pick),
-    // offer only a re-check.
-    return `<div class="lt-yt lt-yt-empty">
-      <div class="lt-yt-msg">${esc(cls.message)}</div>
-      ${checkBtn}
-    </div>`;
-  }
-  // resolved_with_matrix — each release is a pickable rescue target (U5).
-  const staleFlag = cls.stale
-    ? `<div class="lt-yt-stale">${esc(cls.message)}</div>`
-    : '';
-  const targets = youtubeRescueTargets(result);
-  const rows = targets.map((t) => {
-    const bestStr = (t.best_distance != null)
-      ? `dist ${t.best_distance.toFixed(3)}` : 'no distance';
-    const tc = (t.track_count != null) ? `${t.track_count}t` : '';
-    const yr = (t.year != null) ? String(t.year) : '';
-    const meta = [yr, tc, bestStr].filter((x) => x).join(' · ');
-    // The browse id links to the YT Music album in a NEW tab so the operator
-    // can eyeball what it actually is without losing their place in the list
-    // (stopPropagation keeps the click off the row-collapse handler).
-    const ytUrl = youtubeBrowseUrl(t.yt_browse_id);
-    const idHtml = ytUrl
-      ? `<a class="lt-yt-id" href="${esc(ytUrl)}" target="_blank" rel="noopener" onclick="event.stopPropagation()">${esc(t.yt_browse_id)}</a>`
-      : `<span class="lt-yt-id">${esc(t.yt_browse_id)}</span>`;
-    return `<div class="lt-yt-row">
-      ${idHtml}
-      <span class="lt-yt-meta">${esc(meta)}</span>
-      <button class="lt-yt-rescue" type="button" onclick="event.stopPropagation(); window.pickYoutubeRescue(${id}, ${jsArg(t.yt_browse_id)})">Rescue from this</button>
-    </div>`;
-  }).join('');
-  return `<div class="lt-yt lt-yt-matrix">
-    ${staleFlag}
-    <div class="lt-yt-rows">${rows}</div>
-    ${checkBtn}
-  </div>`;
-}
-
-/**
- * Inline "spectral: <grade> ~<kbps>kbps" fragment for the on-disk quality
- * line, coloured by grade (green genuine / red suspect-or-transcode /
- * neutral otherwise — matching `pipeline.js`). Returns '' when the grade is
- * unknown (NULL `current_spectral_grade` — pre-2026-05-17 imports or
- * lossy-source transcodes): "if known". Pure.
- *
- * @param {Object} row  The worklist row.
- * @returns {string}
- */
-function renderSpectralFragment(row) {
-  const grade = row && row.current_spectral_grade
-    ? String(row.current_spectral_grade) : '';
-  if (!grade) return '';
-  const br = (row.current_spectral_bitrate != null)
-    ? ` ~${Number(row.current_spectral_bitrate)}kbps` : '';
-  const color = (grade === 'genuine') ? '#6d6'
-    : (grade === 'suspect' || grade === 'likely_transcode') ? '#d88'
-    : '#caa';
-  return ` · <span style="color:${color};">spectral: ${esc(grade)}${esc(br)}</span>`;
-}
-
-/**
- * Render the full console shell synchronously on open. Each panel starts
- * in its loading state; the independent loaders patch them as their
- * fetches settle. Band-aware emphasis (`consoleEmphasis`) decides which
- * panel leads: `Missing` / unfindable rows lead with why-unfindable;
- * on-disk rows lead with the band-vs-intent header. Pure.
- *
- * @param {Object} row  The worklist row (carries band, source, intent).
- * @param {Object|null} [ytResult]  The row's cached resolver result (from
- *   a prior Check YouTube this session), so a console restore after a list
- *   re-render doesn't wipe a resolved matrix back to `never_run` (#398).
- * @returns {string}
- */
-function renderConsoleShell(row, ytResult) {
-  const id = row.id;
-  const emphasis = consoleEmphasis(row);
-  const leadUnfindable = emphasis.lead === 'unfindable';
-  // Band-vs-intent header for on-disk rows (R8). `band` is the on-disk
-  // band; `target_format` is the request's intent. The console surfaces
-  // the comparison; the exact on-disk codec is in the peers/quality data.
-  // The current spectral measurement (if known) rides along on the same
-  // line so the operator can judge the on-disk quality before deciding to
-  // Set imported / Delete.
-  const band = String(row.band || '').toLowerCase();
-  const intent = row.target_format || (row.min_bitrate ? `${row.min_bitrate}k` : 'default');
-  const bandVsIntent = !leadUnfindable
-    ? `<div class="lt-band-intent lt-panel-lead">
-        <div class="lt-panel-title">Quality vs intent</div>
-        <div class="lt-panel-body">on disk: <strong>${esc(bandLabel(band))}</strong> · intent: <strong>${esc(String(intent))}</strong>${renderSpectralFragment(row)}</div>
-      </div>`
-    : '';
-  const unfindablePanel = renderPanel(
-    'unfindable', id, 'Why unfindable',
-    renderPanelLoading('triage'), leadUnfindable);
-  const peersPanel = renderPanel(
-    'peers', id, 'Soulseek peers seen', renderPanelLoading('peers'), false);
-  const rescuesPanel = renderPanel(
-    'rescues', id, 'Recent rescue attempts', renderPanelLoading('rescues'), false);
-  const siblingsPanel = renderPanel(
-    'siblings', id, 'Sibling pressings', renderPanelLoading('siblings'), false);
-  // The YouTube panel opens in `never_run` — no fetch (U4 must not
-  // auto-call the side-effectful resolver GET) — unless this row already
-  // resolved this session, in which case the cached result renders so a
-  // console restore doesn't discard the operator's matrix.
-  const youtubePanel = renderPanel(
-    'youtube', id, 'YouTube Music', renderYoutubeBody(ytResult || null, id), false);
-  // Order: lead panel first. For Missing/unfindable, why-unfindable leads;
-  // for on-disk rows, the band-vs-intent header leads, then why-unfindable.
-  const ordered = leadUnfindable
-    ? [unfindablePanel, peersPanel, rescuesPanel, siblingsPanel, youtubePanel]
-    : [bandVsIntent, unfindablePanel, peersPanel, rescuesPanel, siblingsPanel, youtubePanel];
-  // The action bar (U6) renders FIRST so the operator's secondary actions
-  // (accept-sibling / set-intent / re-search) stay reachable without
-  // scrolling past the evidence panels. The cohort row carries
-  // `mb_release_group_id` (#398 — projected by `_LONG_TAIL_SELECT`), so the
-  // accept-sibling control renders in its final enabled/disabled state at
-  // open — no detail-fetch stamp, no post-hoc re-render.
-  const actionsBar = renderActionsBar(row);
-  return `<div class="lt-console">${actionsBar}${ordered.join('')}</div>`;
-}
-
-/**
- * Render the secondary-action bar (U6). Pure. Carries four controls:
- *   * Set imported — accept the on-disk quality (wanted → imported) via the
- *     existing `POST /api/pipeline/update` surface.
- *   * Accept-sibling — the Replace operator action, MB-only (KTD7). When
- *     `canAcceptSibling(row, rg)` is false (Discogs-sourced or no MB
- *     release group) the button is disabled with a one-line reason.
- *   * Set-intent — the current intent rendered as a badge with a toggle to
- *     the opposite intent (`intentToggleTarget`).
- *   * Delete request — remove the request entirely via the existing
- *     `POST /api/pipeline/delete` surface (danger-styled).
- *
- * The release-group id is read from the row's `mb_release_group_id`,
- * projected straight off the cohort row by `_LONG_TAIL_SELECT` (#398) — so
- * the bar renders in its final state at console open (no detail-fetch
- * stamp, no post-hoc patch).
- *
- * @param {Object} row  The worklist row (band, source, intent, rg).
- * @returns {string}
- */
-function renderActionsBar(row) {
-  const id = row.id;
-  const rg = row.mb_release_group_id || null;
-  // Discogs / no-rg disable reason (KTD7 — no MB↔Discogs adapter).
-  const acceptReason = acceptDisabledReason(row, rg);
-  const acceptBtn = canAcceptSibling(row, rg)
-    ? `<button class="lt-act-btn lt-act-accept" type="button" onclick="event.stopPropagation(); window.longTailAcceptSibling(${id})">Accept a sibling pressing…</button>`
-    : `<button class="lt-act-btn lt-act-accept" type="button" disabled title="${esc(acceptReason)}">Accept a sibling pressing…</button>
-       <span class="lt-act-note">${esc(acceptReason)}</span>`;
-
-  const intent = overrideToIntent(row.target_format);
-  const intentLabel = intent === 'lossless' ? 'lossless' : 'default';
-  const toggleTarget = intentToggleTarget(row.target_format);
-  const toggleLabel = toggleTarget === 'lossless'
-    ? 'switch to lossless' : 'accept current floor';
-  const intentCtl = `<div class="lt-act-intent">
-      <span class="lt-act-label">intent:</span>
-      <span class="badge lt-intent-badge lt-intent-${esc(intentLabel)}">${esc(intentLabel)}</span>
-      <button class="lt-act-btn lt-act-intent-toggle" type="button" onclick="event.stopPropagation(); window.longTailSetIntent(${id})">${esc(toggleLabel)}</button>
-    </div>`;
-
-  // Triage actions, wired to EXISTING endpoints (no new backend):
-  //   * Set imported → POST /api/pipeline/update {status:"imported"} —
-  //     accept the on-disk quality and stop searching this request.
-  //   * Delete request → POST /api/pipeline/delete — remove it entirely.
-  // Both leave the wanted cohort on success; the handlers drop just that
-  // row from the DOM (no full re-render — keeps scroll position).
-  const importBtn = `<button class="lt-act-btn lt-act-import" type="button" onclick="event.stopPropagation(); window.longTailSetImported(${id})">Set imported</button>`;
-  const deleteBtn = `<button class="lt-act-btn lt-act-delete" type="button" onclick="event.stopPropagation(); window.longTailDeleteRequest(${id})">Delete request</button>`;
-
-  return `<div class="lt-actions" id="lt-actions-${id}">
-    <div class="lt-act-group">${importBtn}</div>
-    <div class="lt-act-group">${acceptBtn}</div>
-    <div class="lt-act-group">${intentCtl}</div>
-    <div class="lt-act-group">${deleteBtn}</div>
-  </div>`;
-}
-
-/**
- * Patch one panel's body in place, guarded by the row's console token so a
- * stale fetch (operator moved on) doesn't paint. DOM-side.
- *
- * @param {number} id     album_requests.id
- * @param {string} name   Panel slug.
- * @param {number} token  The token captured when the console opened.
- * @param {string} html   New body HTML.
- * @returns {void}
- */
-function patchPanel(id, name, token, html) {
-  if (typeof document === 'undefined') return;
-  if (consoleIsStale(consoleStates, id, token)) return;  // stale console — discard.
-  const panel = document.getElementById(`lt-panel-${name}-${id}`);
-  if (!panel) return;
-  const body = panel.querySelector('.lt-panel-body');
-  if (body) body.innerHTML = html;
-}
-
-/**
- * Load the why-unfindable panel from `GET /api/triage/<id>`. Independent —
- * a failure renders only this panel's error affordance.
- *
- * @param {number} id
- * @param {number} token
- * @returns {Promise<void>}
- */
-async function loadUnfindablePanel(id, token) {
-  try {
-    const r = await fetch(`${API}/api/triage/${id}`);
-    if (!r.ok) throw new Error(`HTTP ${r.status}`);
-    const data = await r.json();
-    patchPanel(id, 'unfindable', token, renderUnfindableBody(data));
-  } catch (_e) {
-    patchPanel(id, 'unfindable', token, renderPanelError('triage'));
-  }
-}
-
-/**
- * Load the peers + rescues panels from one `GET /api/pipeline/<id>` fetch.
- * The sibling-pressings panel is fired independently by the console-open
- * path off the cohort row's own `mb_release_group_id` (#398 — the row
- * carries it now; no detail-fetch stamp).
- *
- * @param {number} id
- * @param {number} token
- * @param {boolean} inFlightFlag  The worklist row's `in_flight_rescue`.
- * @returns {Promise<void>}
- */
-async function loadPipelinePanels(id, token, inFlightFlag) {
-  try {
-    const r = await fetch(`${API}/api/pipeline/${id}`);
-    if (!r.ok) throw new Error(`HTTP ${r.status}`);
-    const data = await r.json();
-    patchPanel(id, 'peers', token, renderPeersBody(data.last_search, id));
-    patchPanel(id, 'rescues', token, renderRescuesBody(data.history, inFlightFlag));
-  } catch (_e) {
-    patchPanel(id, 'peers', token, renderPanelError('peers'));
-    patchPanel(id, 'rescues', token, renderPanelError('rescue history'));
-  }
-}
-
-/**
- * Load the sibling-pressings panel from `GET /api/release-group/<rg>`.
- * Independent — a slow / failing release-group renders only this panel's
- * error affordance and never blocks the others. `rgId` doubles as the
- * Discogs master id for Discogs-sourced rows (KTD-1 — numeric ids share
- * this column); the route dispatches numeric ids to the Discogs master
- * endpoint server-side (`web/routes/browse.py::get_release_group`, #501
- * item 1), so a Discogs row with a resolved master gets REAL sibling
- * pressings here, not a doomed MB lookup. Rows with no release group /
- * master at all (masterless Discogs releases, or a legacy MB row with
- * none) render an explicit "no sibling data" state instead of fetching.
- *
- * @param {number} id
- * @param {number} token
- * @param {string|null} rgId  The request's `mb_release_group_id` (MB
- *   release-group UUID or Discogs numeric master id), or null.
- * @returns {Promise<void>}
- */
-async function loadSiblingsPanel(id, token, rgId) {
-  if (!rgId) {
-    patchPanel(id, 'siblings', token,
-      '<div class="lt-panel-empty">No release group — sibling pressings unavailable for this request.</div>');
-    return;
-  }
-  try {
-    const r = await fetch(`${API}/api/release-group/${encodeURIComponent(rgId)}`);
-    if (!r.ok) throw new Error(`HTTP ${r.status}`);
-    const data = await r.json();
-    patchPanel(id, 'siblings', token, renderSiblingsBody(data));
-  } catch (_e) {
-    patchPanel(id, 'siblings', token, renderPanelError('sibling pressings'));
-  }
-}
-
-/**
- * Look up the cached worklist row for an id (band-aware console needs the
- * row's band / source / intent). Returns `null` when the cohort isn't
- * loaded or the id is gone.
- *
- * @param {number} id
- * @returns {Object|null}
- */
-function consoleRow(id) {
-  const rows = Array.isArray(state.longTail.rows) ? state.longTail.rows : [];
-  return rows.find((r) => r && r.id === id) || null;
-}
-
-/**
- * Toggle the in-place action console for one row (U4). Opening renders the
- * console shell synchronously (band-aware), then fires the evidence panels
- * INDEPENDENTLY — each patches its own container as it settles, so a slow
- * or failing panel never blanks the console or drops another panel. A
- * per-row token stamps the fetch so a stale console (operator clicked
- * another row, or re-collapsed this one) is discarded before it paints.
- *
- * The open/closed state is tracked in `consoleStates` (the row's `open`
- * flag) so a list re-render (post-action single-row patch, band switch,
- * search repaint) can restore expanded consoles via
- * {@link restoreLongTailConsoles} instead of collapsing them (#398 / KTD8
- * fidelity).
- *
- * @param {number} id  album_requests.id
- * @returns {void}
- */
-export function toggleLongTailDetail(id) {
-  if (typeof document === 'undefined') return;
-  const el = document.getElementById(`lt-detail-${id}`);
-  if (!el) return;
-  // Re-opening or collapsing this console: drop stale action guards so the
-  // fresh console's buttons aren't silently dead from a prior mid-flight op
-  // (LT-R1). Any outstanding fetch's result is still token-discarded below.
-  consoleClearGuards(consoleStates, id);
-  if (el.classList.contains('open')) {
-    el.classList.remove('open');
-    // Bumps the token (any in-flight panel fetch is discarded on resolve)
-    // and marks the row closed.
-    consoleClose(consoleStates, id);
-    return;
-  }
-  openConsole(id, el);
-}
-
-/**
- * Open one row's console: render the shell, mark it open in
- * `consoleStates`, and fire the independent panel loads. Shared by the
- * row-click toggle and the post-re-render restore path. DOM-side.
- *
- * @param {number} id  album_requests.id
- * @param {HTMLElement} el  The row's `.lt-detail` container.
- * @returns {void}
- */
-function openConsole(id, el) {
-  const token = consoleOpen(consoleStates, id);
-  const row = consoleRow(id) || { id };
-  el.innerHTML = renderConsoleShell(row, consoleYoutubeResult(consoleStates, id));
-  el.classList.add('open');
-  const inFlightFlag = !!row.in_flight_rescue;
-  // Independent panel loads — no Promise.all, no shared await. One panel's
-  // rejection can't reject another's promise (each loader try/catches and
-  // patches its own error affordance). The YouTube panel is already in its
-  // never_run state from the shell (no fetch in U4). Siblings reads the rg
-  // straight off the cohort row (#398 — `_LONG_TAIL_SELECT` projects it).
-  loadUnfindablePanel(id, token);
-  loadPipelinePanels(id, token, inFlightFlag);
-  loadSiblingsPanel(id, token, row.mb_release_group_id || null);
-}
-
-/**
- * Re-open every console the operator had expanded after a list re-render
- * wiped the DOM (#398 / KTD8 fidelity — the post-action single-row patch,
- * a band switch, or a search repaint must not collapse an open console).
- * Rows not currently rendered (filtered out by band/search) are left in
- * the set — switching back to their band restores them. Ids whose rows
- * leave the cohort for good are pruned by `removeRowFromCohort` (surgical
- * removes) and by `loadLongTail`'s fresh-cohort intersect (full refetch).
- * Re-opening re-fires the panel loads: the old DOM (and any fetched panel
- * state) was destroyed with the innerHTML wipe, and the per-row token
- * bump discards any in-flight fetch stamped against the old console.
- * In-flight action guards are deliberately NOT cleared — an outstanding
- * operation keeps its double-fire protection across the re-render.
- *
- * @returns {void}
- */
-export function restoreLongTailConsoles() {
-  if (typeof document === 'undefined') return;
-  for (const id of consoleOpenIds(consoleStates)) {
-    const el = document.getElementById(`lt-detail-${id}`);
-    if (!el || el.classList.contains('open')) continue;
-    openConsole(id, el);
-  }
-}
-
-// --- Console rescue flow (U5) ----------------------------------------
-//
-// Two steps: "Check YouTube" runs the slow side-effectful resolver GET and
-// re-renders the YouTube panel with pickable rescue targets; picking a
-// target opens a confirm overlay and submits the rescue. Both steps are
-// double-fire-guarded (a module-scoped Set keyed by request id). The
-// resolver result is stamped with the row's console token so a stale result
-// (operator moved on) is discarded before it paints.
-
-/**
- * @typedef {Object} RescueCopy
- * @property {string} title    Short headline for the outcome.
- * @property {string} detail   One-line operator-facing explanation.
- * @property {'success'|'error'} tone  Drives toast styling (error → red).
- */
-
-/**
- * Map a YouTube-rescue ingest outcome to specific operator-facing console
- * copy. Pure / DOM-free. Keys mirror the EXACT outcome vocabulary in
- * `lib/youtube_ingest_service.py` / `web/routes/youtube.py`
- * (`OUTCOME_HTTP_STATUS`): `accepted`, `request_not_found`, `wrong_state`,
- * `in_flight`, `no_resolver_mapping`, `track_count_precheck_failed`,
- * `transient`.
- *
- * `result` is the parsed rescue-submit response body — its `outcome`
- * selects the copy; `download_log_id` / `detail` decorate the `in_flight`
- * and `track_count_precheck_failed` messages with the specifics the
- * backend returns. An unknown outcome falls back to a generic error so a
- * future backend value never renders blank.
- *
- * @param {{outcome?: string, download_log_id?: number|null, detail?: string|null, error?: string|null}|null|undefined} result
- * @returns {RescueCopy}
- */
-export function rescueOutcomeCopy(result) {
-  const outcome = String((result && result.outcome) || '');
-  const detail = (result && result.detail) ? String(result.detail) : '';
-  const logId = (result && result.download_log_id != null)
-    ? result.download_log_id : null;
-  switch (outcome) {
-    case 'accepted':
-      return {
-        title: 'Rescue queued',
-        detail: logId != null
-          ? `Rescue queued (download_log #${logId}). The importer owns the import; the row updates when it lands.`
-          : 'Rescue queued. The importer owns the import; the row updates when it lands.',
-        tone: 'success',
-      };
-    case 'in_flight':
-      return {
-        title: 'Rescue already running',
-        detail: logId != null
-          ? `A rescue is already running for this request (download_log #${logId}).`
-          : 'A rescue is already running for this request.',
-        tone: 'error',
-      };
-    case 'wrong_state':
-      return {
-        title: 'Request changed',
-        detail: 'This request is no longer wanted/manual — refresh and try again.',
-        tone: 'error',
-      };
-    case 'no_resolver_mapping':
-      return {
-        title: 'No resolver mapping',
-        detail: 'No cached YouTube mapping for this release — re-run Check YouTube first.',
-        tone: 'error',
-      };
-    case 'track_count_precheck_failed':
-      return {
-        title: 'Track-count mismatch',
-        detail: detail
-          ? `Track-count precheck failed: ${detail}`
-          : 'Track-count precheck failed: the resolver and MB mirror disagree — refresh and re-check.',
-        tone: 'error',
-      };
-    case 'transient':
-      return {
-        title: 'Temporary failure',
-        detail: detail
-          ? `Temporary failure: ${detail}. Retry.`
-          : 'Temporary failure (DB / mirror hiccup). Retry.',
-        tone: 'error',
-      };
-    case 'request_not_found':
-      return {
-        title: 'Request not found',
-        detail: 'This request no longer exists — refresh.',
-        tone: 'error',
-      };
-    default:
-      return {
-        title: 'Rescue failed',
-        detail: detail || (result && result.error
-          ? String(result.error)
-          : `Unexpected outcome: ${outcome || 'unknown'}.`),
-        tone: 'error',
-      };
-  }
-}
-
-/**
- * True while a rescue confirm overlay is open. The overlay mounts into the
- * single shared modal host, so a second `pickYoutubeRescue` would overwrite
- * the first overlay's DOM and orphan its promise — this boolean serialises
- * them (one confirm at a time).
- */
-let rescueConfirmOpen = false;
-
-/**
- * Re-render just the YouTube panel body for one row's open console, guarded
- * by the row's console token so a stale result doesn't paint. DOM-side.
- *
- * @param {number} id
- * @param {number} token  The console token captured when the resolve fired.
- * @param {{outcome?: string, youtube_releases?: Array<Object>|null, from_cache?: boolean, error_message?: string|null}|null} result
- * @returns {void}
- */
-function patchYoutubePanel(id, token, result) {
-  patchPanel(id, 'youtube', token, renderYoutubeBody(result, id));
-}
-
-/**
- * "Check YouTube" handler (U5) — replaces U4's placeholder toast. Runs the
- * slow, side-effectful resolver GET for the row's `mb_release_id`, then
- * re-renders the YouTube panel with the fresh classification.
- *
- * Guards:
- *   * Double-fire — `consoleCanStart(consoleStates, id, 'resolve')`; a
- *     second click while outstanding returns immediately.
- *   * Disabled button — the live "Check YouTube" / "Retry" button is
- *     disabled + relabelled while outstanding so the operator sees progress
- *     and can't re-click it.
- *
- * The settled result is cached via {@link consoleSetYoutubeResult} and
- * painted against the row's CURRENT console token ({@link consoleToken} —
- * not the one captured at fire time): the panel container is per-row, so
- * the only console it can paint into is this row's — including one
- * re-created by a #398 restore while the fetch was outstanding. A
- * collapsed console's hidden DOM may be touched; harmless, since reopening
- * re-renders the shell (from the cache).
- *
- * The resolver identifier is the request's `mb_release_id` (an MB release
- * MBID or a Discogs release id) — the same id the resolver's
- * `?identifier=` query takes. A row without one cannot be resolved; the
- * panel shows that explicitly rather than firing a doomed fetch.
- *
- * @param {number} id  album_requests.id
- * @returns {Promise<void>}
- */
-export async function checkYoutube(id) {
-  if (!consoleCanStart(consoleStates, id, 'resolve')) return;  // double-fire guard.
-  const row = consoleRow(id);
-  const identifier = row && row.mb_release_id ? String(row.mb_release_id) : '';
-  // Paint against the row's CURRENT token, re-read at paint time (the
-  // resolver-settle semantic — see consoleToken's docstring): the youtube
-  // panel container is per-row, so any console this can reach is this
-  // row's — including one re-created by a #398 restore mid-fetch.
-  const currentToken = () => consoleToken(consoleStates, id);
-  if (!identifier) {
-    patchYoutubePanel(id, currentToken(), /** @type {any} */ (
-      { outcome: 'transient', error_message: 'No release identifier on this request.' }));
-    consoleSettle(consoleStates, id, 'resolve');
-    return;
-  }
-  setYoutubeChecking(id);
-  try {
-    const r = await fetch(
-      `${API}/api/youtube-album?identifier=${encodeURIComponent(identifier)}`);
-    // 404/503 still carry a typed body; the classifier maps any non-`ok`
-    // outcome to `resolver_failed`, so we read the body regardless of
-    // status and only fall back to a synthetic failure if the body is
-    // unreadable.
-    let result;
-    try {
-      result = await r.json();
-    } catch (_e) {
-      result = { outcome: 'transient', error_message: `HTTP ${r.status}` };
-    }
-    consoleSetYoutubeResult(consoleStates, id, result);
-    patchYoutubePanel(id, currentToken(), result);
-  } catch (_e) {
-    patchYoutubePanel(id, currentToken(), /** @type {any} */ (
-      { outcome: 'transient', error_message: 'Could not reach the resolver. Retry.' }));
-  } finally {
-    consoleSettle(consoleStates, id, 'resolve');
-  }
-}
-
-/**
- * Swap the YouTube panel into an in-progress state while the resolver GET
- * is outstanding (disabled button + spinner copy). DOM-side; no-op when the
- * panel isn't mounted (Node tests / collapsed console).
- *
- * @param {number} id
- * @returns {void}
- */
-function setYoutubeChecking(id) {
-  if (typeof document === 'undefined') return;
-  const panel = document.getElementById(`lt-panel-youtube-${id}`);
-  if (!panel) return;
-  const body = panel.querySelector('.lt-panel-body');
-  if (!body) return;
-  body.innerHTML = `<div class="lt-yt lt-yt-checking">
-    <div class="lt-yt-msg">Resolving against YouTube Music…</div>
-    <button class="lt-yt-check" type="button" disabled>Checking…</button>
-  </div>`;
-}
-
-/**
- * Render the rescue confirm dialog body. Pure. Mirrors the
- * `.confirm-box` shell used by `replace_picker.js` so the visual language
- * matches the only other destructive-confirm in the app.
- *
- * @param {number} id        album_requests.id
- * @param {string} browseId  The YT Music album browseId being submitted.
- * @param {Object|null} row  The worklist row (for the album label).
- * @returns {string}
- */
-export function renderRescueConfirm(id, browseId, row) {
-  const album = (row && row.album_title) ? String(row.album_title) : `request #${id}`;
-  const artist = (row && row.artist_name) ? String(row.artist_name) : '';
-  const label = artist ? `${artist} — ${album}` : album;
-  return `<div class="confirm-box" role="dialog" aria-modal="true">
-    <h3>Rescue from YouTube Music?</h3>
-    <p>Queue a YouTube-Music rescue for:<br><strong>${esc(label)}</strong></p>
-    <p>Target album:<br><code>${esc(browseId)}</code></p>
-    <p style="font-size:0.85em;color:#999;">The request stays <code>wanted</code> until the
-    importer lands the rescue (minutes later) — the row won't move bands immediately.</p>
-    <div class="actions">
-      <button class="btn" id="lt-rescue-cancel">Cancel</button>
-      <button class="btn p-btn" id="lt-rescue-confirm">Rescue</button>
-    </div>
-  </div>`;
-}
-
-/**
- * The dedicated mount node for the rescue confirm overlay. Reuses the
- * shared `replace-picker-modal` host (same pattern as the Replace picker —
- * one modal host, wiped on close).
- *
- * @returns {HTMLElement|null}
- */
-function rescueModalHost() {
-  if (typeof document === 'undefined') return null;
-  return document.getElementById('replace-picker-modal');
-}
-
-/**
- * Open the Promise-based rescue confirm overlay. Mirrors
- * `replace_picker.js`'s overlay shell: a full-screen `.confirm-overlay`
- * backdrop (click-to-cancel) wrapping the `.confirm-box`. Resolves `true`
- * on confirm, `false` on cancel / backdrop click. DOM-side.
- *
- * @param {number} id
- * @param {string} browseId
- * @param {Object|null} row
- * @returns {Promise<boolean>}
- */
-function confirmRescue(id, browseId, row) {
-  const host = rescueModalHost();
-  if (!host) return Promise.resolve(false);
-  return new Promise((resolve) => {
-    let settled = false;
-    /** @param {boolean} ok */
-    function close(ok) {
-      if (settled) return;
-      settled = true;
-      host.style.display = 'none';
-      host.innerHTML = '';
-      resolve(ok);
-    }
-    host.innerHTML = `<div class="confirm-overlay">${renderRescueConfirm(id, browseId, row)}</div>`;
-    host.style.display = '';
-    const overlay = host.querySelector('.confirm-overlay');
-    if (overlay) {
-      overlay.addEventListener('click', (event) => {
-        if (event.target === overlay) close(false);  // backdrop-click cancel.
-      });
-    }
-    const cancel = host.querySelector('#lt-rescue-cancel');
-    if (cancel) cancel.addEventListener('click', () => close(false));
-    const confirm = host.querySelector('#lt-rescue-confirm');
-    if (confirm) confirm.addEventListener('click', () => close(true));
-  });
-}
-
-/**
- * Pick a rescue target (U5) — opens the confirm overlay, and on confirm
- * submits the rescue. Double-fire-guarded on the submit step.
- *
- * @param {number} id        album_requests.id
- * @param {string} browseId  The chosen target's `yt_browse_id`.
- * @returns {Promise<void>}
- */
-export async function pickYoutubeRescue(id, browseId) {
-  if (rescueConfirmOpen) return;  // one confirm overlay at a time (shared host).
-  rescueConfirmOpen = true;
-  try {
-    const row = consoleRow(id);
-    const ok = await confirmRescue(id, browseId, row);
-    if (!ok) return;
-    await submitYoutubeRescue(id, browseId);
-  } finally {
-    rescueConfirmOpen = false;
-  }
-}
-
-/**
- * Submit the rescue (`POST /api/pipeline/<id>/youtube-rescue {browse_id}`)
- * and map the outcome to a toast. The submit is fire-and-forget from the
- * UI's perspective: the importer owns the band transition (KTD4) minutes
- * later, so on `accepted` we deliberately do NOT refetch or re-render —
- * that would collapse every open console and lose the operator's scroll
- * position (operator preference). Double-fire-guarded.
- *
- * @param {number} id
- * @param {string} browseId
- * @returns {Promise<void>}
- */
-async function submitYoutubeRescue(id, browseId) {
-  if (!consoleCanStart(consoleStates, id, 'submit')) return;  // double-fire guard.
-  try {
-    let result;
-    try {
-      const r = await fetch(`${API}/api/pipeline/${id}/youtube-rescue`, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ browse_id: browseId }),
-      });
-      result = await r.json();
-    } catch (_e) {
-      result = { outcome: 'transient', error_message: 'Submit failed — retry.' };
-    }
-    const copy = rescueOutcomeCopy(result);
-    if (typeof toast === 'function') {
-      toast(`${copy.title}: ${copy.detail}`, copy.tone === 'error');
-    }
-    // Toast and nothing else (operator preference): the rescue stays in
-    // flight for minutes and the importer owns the band transition, so a
-    // refetch/re-render here only collapses every open console and throws
-    // away the operator's scroll position. The "rescue running" badge
-    // reconciles on the next manual Refresh.
-  } finally {
-    consoleSettle(consoleStates, id, 'submit');
-  }
-}
-
-/**
- * Single-row refetch + patch (KTD8). Fetches just this request's
- * authoritative banded row via `GET /api/pipeline/long-tail?id=<id>` and
- * replaces it in `state.longTail.rows`, then re-renders the list. NO
- * full-cohort re-band (the heaviest read in the app), NO optimistic band
- * move — the row stays in `Missing` until the importer completes (KTD4).
- *
- * A 404 (the row left the `wanted` worklist — e.g. already imported)
- * removes it from the cohort. A failed fetch is non-fatal: the local
- * in-flight mark from {@link markRowInFlight} already gives the operator a
- * signal; the next Refresh reconciles.
- *
- * The re-render does NOT collapse the acted-on row's console (or any
- * other): `renderPipeline`'s long-tail paint ends with
- * {@link restoreLongTailConsoles}, which re-opens every console marked
- * open in `consoleStates` against the fresh DOM (#398).
- *
- * @param {number} id
- * @returns {Promise<void>}
- */
-async function refetchLongTailRow(id) {
-  let data;
-  try {
-    const r = await fetch(`${API}/api/pipeline/long-tail?id=${encodeURIComponent(String(id))}`);
-    if (r.status === 404) {
-      removeRowFromCohort(id);
-      renderLongTail();
-      return;
-    }
-    if (!r.ok) return;
-    data = await r.json();
-  } catch (_e) {
-    return;  // non-fatal — local in-flight mark stands; Refresh reconciles.
-  }
-  const fresh = data && data.result;
-  if (!fresh) return;
-  patchRowInCohort(id, fresh);
-  renderLongTail();
-}
-
-/**
- * Replace one row in the cohort with a freshly-refetched authoritative row.
- * Pure-ish: mutates `state.longTail.rows`. No-op when the cohort isn't
- * loaded or the id is gone (a concurrent full refetch already reconciled).
- *
- * @param {number} id
- * @param {Object} fresh  The refetched `LongTailRow`.
- * @returns {void}
- */
-function patchRowInCohort(id, fresh) {
-  const rows = Array.isArray(state.longTail.rows) ? state.longTail.rows : null;
-  if (!rows) return;
-  const idx = rows.findIndex((r) => r && r.id === id);
-  if (idx === -1) return;
-  rows[idx] = fresh;
-}
-
-/**
- * Drop one row from the cohort (the single-row refetch 404'd — the row left
- * the `wanted` worklist). Pure-ish: mutates `state.longTail.rows`.
- *
- * @param {number} id
- * @returns {void}
- */
-function removeRowFromCohort(id) {
-  // The row is gone for good — prune its ENTIRE console state (open flag,
-  // token, cached resolver result, in-flight guards) in one call. One map,
-  // one prune function (#481 item 1) — the single-row drop.
-  consolePrune(consoleStates, (rowId) => rowId !== id);
-  const rows = Array.isArray(state.longTail.rows) ? state.longTail.rows : null;
-  if (!rows) return;
-  const idx = rows.findIndex((r) => r && r.id === id);
-  if (idx !== -1) rows.splice(idx, 1);
-}
-
-/**
- * Remove one row's DOM nodes in place — the `.lt-item` and its trailing
- * `.lt-detail` console — WITHOUT re-rendering the list. Keeps the operator's
- * scroll position and every other open console intact (the whole point of
- * the Set imported / Delete triage flow). Pair with `removeRowFromCohort`
- * to keep `state.longTail.rows` consistent; the band-tab counts reconcile on
- * the next render (band switch / Refresh). DOM-side; no-op when not mounted.
- *
- * @param {number} id
- * @returns {void}
- */
-function removeRowElement(id) {
-  if (typeof document === 'undefined') return;
-  const detail = document.getElementById(`lt-detail-${id}`);
-  if (!detail) return;
-  const item = detail.previousElementSibling;
-  detail.remove();
-  if (item && item.classList && item.classList.contains('lt-item')) {
-    item.remove();
-  }
-}
-
-// --- Console secondary actions (U6) ----------------------------------
-//
-// Operator actions wired to EXISTING endpoints — no parallel code
-// paths (R16), no new backend:
-//   * set-imported   → `POST /api/pipeline/update {id, status:"imported"}`
-//   * accept-sibling → `openReplacePicker` (standard mode) → `POST .../replace`
-//   * set-intent     → `POST /api/pipeline/set-intent {id, intent}`
-//   * delete-request → `POST /api/pipeline/delete {id}`
-// Freshness (KTD8): set-intent refetches just the acted-on row and patches
-// it (`refetchLongTailRow`). Set-imported / Delete remove just that row's
-// DOM nodes (it leaves the cohort) without a full re-render — keeping the
-// operator's scroll position. The confirmed Replace closes the console and
-// full-cohort refetches.
-
-/**
- * Can the operator accept a sibling pressing for this row? Pure / DOM-free.
- *
- * MB-only in v1 (KTD7 — sibling enumeration is MB-keyed via the release
- * group; there is no MB↔Discogs adapter). `false` when:
- *   * the request is Discogs-sourced (`row.source === 'discogs'`), OR
- *   * there is no `mb_release_group_id` to enumerate siblings from.
- * `true` otherwise (an MB-sourced request with a resolved release group).
- *
- * @param {Object|null|undefined} row  The worklist row.
- * @param {string|null|undefined} releaseGroupId  The resolved MB rg id.
- * @returns {boolean}
- */
-export function canAcceptSibling(row, releaseGroupId) {
-  if (!row || typeof row !== 'object') return false;
-  const source = String(row.source || '').toLowerCase();
-  if (source === 'discogs') return false;
-  if (!releaseGroupId) return false;
-  return true;
-}
-
-/**
- * One-line reason the accept-sibling action is disabled. Pure / DOM-free.
- * Returns the empty string when the action is enabled.
- *
- * @param {Object|null|undefined} row
- * @param {string|null|undefined} releaseGroupId
- * @returns {string}
- */
-export function acceptDisabledReason(row, releaseGroupId) {
-  if (canAcceptSibling(row, releaseGroupId)) return '';
-  const source = String((row && row.source) || '').toLowerCase();
-  if (source === 'discogs') {
-    return 'Sibling pressings are MusicBrainz-only — Discogs requests have no MB release group.';
-  }
-  return 'No MusicBrainz release group for this request — no siblings to pick from.';
-}
-
-/**
- * The intent to POST when toggling. Pure / DOM-free. The set-intent
- * surface is a two-value toggle (`lossless` ⇄ `default`); given the row's
- * current `target_format`, this returns the OPPOSITE intent. A lossless
- * `target_format` (`"lossless"` / `"flac"`) currently means "keep grinding
- * to lossless", so the toggle target is `"default"` (accept current
- * floor); anything else toggles to `"lossless"`.
- *
- * @param {string|null|undefined} targetFormat  The row's `target_format`.
- * @returns {'lossless'|'default'}
- */
-export function intentToggleTarget(targetFormat) {
-  return overrideToIntent(targetFormat) === 'lossless' ? 'default' : 'lossless';
-}
-
-/**
- * Build the standard-mode `openReplacePicker` options for one cohort row.
- * Pure / DOM-free. `sourceLabel` is the "Artist — Album" the picker shows
- * as the current request; `releaseGroupId` is the stamped MB rg (may be
- * null — the picker lazily resolves a null rg via `resolve-rg`, but U6
- * only enables the button when one is present, KTD7).
- *
- * @param {Object} row
- * @returns {import('./replace_picker.js').ReplacePickerOptionsStandard}
- */
-export function buildAcceptSiblingOptions(row) {
-  const artist = (row && row.artist_name) ? String(row.artist_name) : '';
-  const album = (row && row.album_title) ? String(row.album_title) : '';
-  const label = artist ? `${artist} — ${album}` : (album || `request #${row.id}`);
-  return {
-    sourceRequestId: row.id,
-    releaseGroupId: (row && row.mb_release_group_id) || null,
-    sourceLabel: label,
-    source: 'long-tail',
-  };
-}
-
-/**
- * Accept-a-sibling-pressing handler (U6). Opens the existing Replace picker
- * (standard mode) for the row's MB release group; the picker owns the
- * sibling list + confirm + `POST .../replace`. On a CONFIRMED Replace the
- * old request becomes a frozen `replaced` row that LEAVES the worklist, so
- * we close the console and do a FULL-cohort refetch (KTD8 exception) — not
- * the single-row patch. Cancel / failure is a no-op (the picker toasts its
- * own errors).
- *
- * Reuses the raw `openReplacePicker` (not the `window.openReplacePicker`
- * wrapper) because the wrapper refetches the Pipeline/Manual tabs, not the
- * long-tail cohort — and it doesn't return the confirmed result we need to
- * decide on the full-cohort refetch.
- *
- * @param {number} id  album_requests.id
- * @returns {Promise<void>}
- */
-export async function longTailAcceptSibling(id) {
-  const row = consoleRow(id);
-  if (!row) return;
-  if (!canAcceptSibling(row, row.mb_release_group_id || null)) {
-    // Defensive — the button is disabled in this case, but a stale onclick
-    // (rg dropped between render and click) should fail loudly, not POST.
-    if (typeof toast === 'function') {
-      toast(acceptDisabledReason(row, row.mb_release_group_id || null), true);
-    }
-    return;
-  }
-  const result = await openReplacePicker(buildAcceptSiblingOptions(row));
-  if (!result || result.outcome !== 'confirmed') return;  // cancelled.
-  const resp = result.response || { status: 0, body: {} };
-  const status = resp.status;
-  const body = resp.body || {};
-  if (status === 200) {
-    if (typeof toast === 'function') {
-      const newId = body.new_request_id;
-      toast(newId != null
-        ? `Replaced — new request #${newId}. The old request is frozen for audit.`
-        : 'Replaced — the old request is frozen for audit.', false);
-    }
-    // The old row left the cohort (now `status='replaced'`). Close the
-    // console and FULL-cohort refetch (KTD8 exception — the single-row
-    // patch can't express a row leaving).
-    closeConsole(id);
-    await loadLongTail();
-    return;
-  }
-  // Non-200 confirmed POST — surface the picker's error shape (mirrors
-  // main.js::openReplacePickerAndHandle), leave the cohort untouched.
-  if (typeof toast === 'function') {
-    if (status === 409 && body && body.descendant_request_id) {
-      toast(`Already replaced. New request is #${body.descendant_request_id}.`, true);
-    } else if (body && body.error) {
-      toast(`Replace failed: ${body.error}`, true);
-    } else {
-      toast(`Replace failed (HTTP ${status})`, true);
-    }
-  }
-}
-
-/**
- * Close one row's console (collapse `.lt-detail`, bump its token so an
- * in-flight panel fetch is discarded on resolve). DOM-side; no-op when the
- * console isn't mounted.
- *
- * @param {number} id
- * @returns {void}
- */
-function closeConsole(id) {
-  if (typeof document === 'undefined') return;
-  const el = document.getElementById(`lt-detail-${id}`);
-  if (el) el.classList.remove('open');
-  consoleClose(consoleStates, id);
-}
-
-/**
- * Set-intent handler (U6). Toggles the request's quality intent (lossless
- * ⇄ default) via the EXISTING set-intent surface
- * (`POST /api/pipeline/set-intent {id, intent}` — mirrors
- * `library.js::setIntent`), then single-row refetch-and-patch so the badge
- * reflects the new intent (KTD8). The target intent is the opposite of the
- * row's current `target_format` (`intentToggleTarget`).
- *
- * @param {number} id  album_requests.id
- * @returns {Promise<void>}
- */
-export async function longTailSetIntent(id) {
-  const row = consoleRow(id);
-  if (!row) return;
-  if (!consoleCanStart(consoleStates, id, 'intent')) return;  // double-fire guard.
-  const intent = intentToggleTarget(row.target_format);
-  try {
-    let data;
-    try {
-      const r = await fetch(`${API}/api/pipeline/set-intent`, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ id, intent }),
-      });
-      data = await r.json();
-    } catch (_e) {
-      if (typeof toast === 'function') toast('Failed to set intent', true);
-      return;
-    }
-    if (!data || data.status !== 'ok') {
-      if (typeof toast === 'function') {
-        toast((data && data.error) || 'Failed to set intent', true);
-      }
-      return;
-    }
-    if (typeof toast === 'function') {
-      toast(data.requeued ? `Intent: ${intent} (requeued)` : `Intent: ${intent}`, false);
-    }
-    // Single-row refetch-and-patch (KTD8) — the badge reflects the new intent
-    // off the authoritative refetched row.
-    await refetchLongTailRow(id);
-  } finally {
-    consoleSettle(consoleStates, id, 'intent');
-  }
-}
-
-/**
- * "Set imported" handler (U6) — accept the on-disk quality. Flips the
- * request `wanted → imported` via the EXISTING `POST /api/pipeline/update`
- * surface (same endpoint the Pipeline tab uses), then drops just this row
- * from the cohort: the row leaves the `wanted` worklist, but we remove only
- * its DOM nodes (no full re-render — keeps the operator's scroll position
- * and other open consoles). Double-fire-guarded.
- *
- * @param {number} id  album_requests.id
- * @returns {Promise<void>}
- */
-export async function longTailSetImported(id) {
-  if (!consoleCanStart(consoleStates, id, 'import')) return;  // double-fire guard.
-  try {
-    let data;
-    try {
-      const r = await fetch(`${API}/api/pipeline/update`, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ id, status: 'imported' }),
-      });
-      data = await r.json();
-    } catch (_e) {
-      if (typeof toast === 'function') toast('Failed to set imported', true);
-      return;
-    }
-    if (!data || data.status !== 'ok') {
-      if (typeof toast === 'function') {
-        toast((data && data.error) || 'Failed to set imported', true);
-      }
-      return;
-    }
-    if (typeof toast === 'function') toast(`#${id} set to imported`, false);
-    removeRowFromCohort(id);
-    removeRowElement(id);
-  } finally {
-    consoleSettle(consoleStates, id, 'import');
-  }
-}
-
-/**
- * "Delete request" handler (U6) — remove the request entirely via the
- * EXISTING `POST /api/pipeline/delete` surface (hard delete; content-
- * addressed evidence survives). Confirms first (destructive, irreversible).
- * On success drops just this row's DOM nodes (no full re-render). A 409
- * means a superseding Replace blocks deletion — surfaces the descendant ids
- * so the operator deletes those first. Double-fire-guarded.
- *
- * @param {number} id  album_requests.id
- * @returns {Promise<void>}
- */
-export async function longTailDeleteRequest(id) {
-  if (typeof confirm === 'function'
-      && !confirm(`Delete request #${id}? This removes the wanted request entirely.`)) {
-    return;
-  }
-  if (!consoleCanStart(consoleStates, id, 'delete')) return;  // double-fire guard.
-  try {
-    let status = 0;
-    let data;
-    try {
-      const r = await fetch(`${API}/api/pipeline/delete`, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ id }),
-      });
-      status = r.status;
-      data = await r.json();
-    } catch (_e) {
-      if (typeof toast === 'function') toast('Failed to delete request', true);
-      return;
-    }
-    if (status === 200 && data && data.status === 'ok') {
-      if (typeof toast === 'function') toast(`#${id} deleted`, false);
-      removeRowFromCohort(id);
-      removeRowElement(id);
-      return;
-    }
-    if (status === 409 && data && Array.isArray(data.descendant_request_ids)) {
-      const ids = data.descendant_request_ids.map((d) => `#${d}`).join(', ');
-      if (typeof toast === 'function') {
-        toast(`Can't delete #${id}: superseded by ${ids} — delete those first.`, true);
-      }
-      return;
-    }
-    if (typeof toast === 'function') {
-      toast((data && data.error) || `Delete failed (HTTP ${status})`, true);
-    }
-  } finally {
-    consoleSettle(consoleStates, id, 'delete');
-  }
-}
 
 export const __test__ = {
   MISSING_BAND,
@@ -2395,42 +535,4 @@ export const __test__ = {
   countOtherBandMatches,
   renderLongTailRow,
   renderLongTailBody,
-  // U4 — console pure render helpers + re-exported util classifiers.
-  youtubeSectionState,
-  consoleEmphasis,
-  renderConsoleShell,
-  renderUnfindableBody,
-  renderPeersBody,
-  renderRescuesBody,
-  renderSiblingsBody,
-  renderYoutubeBody,
-  renderPanelError,
-  youtubeHistoryRows,
-  youtubeFailureReason,
-  PEERS_VISIBLE_CAP,
-  // U5 — two-step rescue flow pure helpers.
-  youtubeBestDistance,
-  youtubeRescueTargets,
-  rescueOutcomeCopy,
-  renderRescueConfirm,
-  // #481 item 1 — the console-state Map + its five pure transition helpers.
-  consoleStates,
-  consoleOpen,
-  consoleClose,
-  consolePrune,
-  consoleCanStart,
-  consoleSettle,
-  consoleClearGuards,
-  consoleToken,
-  consoleIsStale,
-  consoleIsOpen,
-  consoleYoutubeResult,
-  consoleSetYoutubeResult,
-  consoleOpenIds,
-  // U6 — secondary action pure helpers.
-  canAcceptSibling,
-  acceptDisabledReason,
-  intentToggleTarget,
-  buildAcceptSiblingOptions,
-  renderActionsBar,
 };
